+++ piyavkin.anton/T1/isoscale.hpp
@@ -1,0 +1,6 @@
+#ifndef ISOSCALE_HPP
+#define ISOSCALE_HPP
+#include "base-types.hpp"
+#include "shape.hpp"
+void isoScale(Shape * shape, point_t center, double k);
+#endif
